pyproject.toml
[project]
name = "astryx-ai-microservice"
version = "0.1.0"
description = "microservice in Python for astryx-backend providing Azure OpenAI API and Tools API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core FastAPI and web server
    "fastapi==0.115.13",
    "uvicorn==0.34.2",
    "python-multipart==0.0.6",
    # Langchain and related libraries
    "langchain==0.3.25",
    "langchain-core==0.3.65",
    "langchain-openai==0.3.16",
    "langgraph==0.5.1",
    "langserve==0.3.1",
    "openai==1.76.2",
    # Configuration and environment management
    "pydantic==2.11.4",
    "pydantic-settings==2.9.1",
    "python-dotenv==1.1.0",
    # Utilities
    "tiktoken==0.9.0",
    "sse-starlette>=2.3.6",
]
